pyproject.toml
[project]
name = "genaiproject"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dotenv>=0.9.9",
    "google-genai>=1.39.1",
    "langchain>=0.3.27",
    "langchain-google-genai>=2.1.12",
    "langchain-openai>=0.3.33",
    "langchain-qdrant>=0.2.1",
    "langgraph>=0.6.7",
    "plotly>=6.3.0",
    "prophet>=1.1.7",
    "seaborn>=0.13.2",
    "streamlit>=1.49.1",
]
